1.4 major revision
Added convenience methods:

userAgent: a manual call to set the video useragent setting. Normally, it's done on init() - but in some cases you want to be able to set it even before running the initialization. (i.e. if you want to inject/add special behaviors on the video BEFORE actually initializing it)

mute & unmute: manually force/set audio state

isPlaying: returns a boolean - if you want to know if the movie has started playing (is still true when paused) basically for knowing if the video is done or has started.

isMuted: returns a boolean - if you want to know the video's current audio state

reflow: you can manually call the re-calculation of all controller positions (useful if you want to dynamically change the size of the video and re-calibrate everything without re-initializing

uniquereplay: on video end, display the play button again  instead of the circular replay button (some people prefer this)

Changed listener callbacks to be more "semantic"

timeupdate -> progress (as in fire everytime the video "progresses" in playing)

progress -> loading (as in the FILE is currently loading/buffering)